Move BaseCollection to data package
Prepare for using it in other packages as well.
